requirements.txt
datasets
torchmetrics
tokenizers
torch-lr-finder
lightning
torch_tb_profiler
pathlib